requirements.txt
pygame
colorama
Pillow
tqdm
pydub
ffmpeg
ffmpeg-python